Add files via upload
It is a desktop based application for hospital which allow user to book appointment by themself and know information about blood bank, Doctors and labs of hospital. The main aim of developing “MEDICURE SYSTEM” is to computerize the Front Office Management of Hospital.
Designed in hibernate and java as front-end and MySQL as back-end.